[Decomp][Player] Cohesion split: extract group-membership methods to PlayerGroup.cpp
